+++ src/modules/Ultrastar/ultrastar_converter.py
@@ -48,30 +48,25 @@ def ultrastar_note_to_midi_note(ultrastar_note: int) -> int:
     return midi_note
 
 
-def get_start_time_from_ultrastar(ultrastar_class: UltrastarTxtValue, pos: int) -> float:
+def get_start_time_from_ultrastar(
+    ultrastar_class: UltrastarTxtValue, pos: int
+) -> float:
     """Calculates the start time from the Ultrastar txt"""
 
-    gap = int(ultrastar_class.gap) / 1000
-    real_bpm = ultrastar_bpm_to_real_bpm(
-        float(ultrastar_class.bpm.replace(",", "."))
-    )
-    start_time = (
-        beat_to_second(int(ultrastar_class.startBeat[pos]), real_bpm) + gap
-    )
+    gap = int(float(ultrastar_class.gap.replace(",", ".")) / 1000)
+    real_bpm = ultrastar_bpm_to_real_bpm(float(ultrastar_class.bpm.replace(",", ".")))
+    start_time = beat_to_second(int(ultrastar_class.startBeat[pos]), real_bpm) + gap
     return start_time
 
 
 def get_end_time_from_ultrastar(ultrastar_class: UltrastarTxtValue, pos: int) -> float:
     """Calculates the end time from the Ultrastar txt"""
 
-    gap = int(ultrastar_class.gap) / 1000
-    real_bpm = ultrastar_bpm_to_real_bpm(
-        float(ultrastar_class.bpm.replace(",", "."))
-    )
+    gap = int(float(ultrastar_class.gap.replace(",", ".")) / 1000)
+    real_bpm = ultrastar_bpm_to_real_bpm(float(ultrastar_class.bpm.replace(",", ".")))
     end_time = (
         beat_to_second(
-            int(ultrastar_class.startBeat[pos])
-            + int(ultrastar_class.durations[pos]),
+            int(ultrastar_class.startBeat[pos]) + int(ultrastar_class.durations[pos]),
             real_bpm,
         )
         + gap