+++ src/modules/Ultrastar/ultrastar_score_calculator.py
@@ -1,4 +1,7 @@
 """Ultrastar score calculator."""
+from dataclasses import dataclass
+
+from dataclasses_json import dataclass_json
 
 import librosa
 
@@ -48,6 +51,8 @@ def add_point(note_type: str, points: Points) -> Points:
     return points
 
 
+@dataclass_json
+@dataclass
 class Score:
     """Docstring"""
 
@@ -121,8 +126,10 @@ def calculate_score(pitched_data: PitchedData, ultrastar_class: UltrastarTxtValu
         for part in range(parts):
             start = start_time + step_size * part
             end = start + step_size
+
             if end_time < end or part == parts - 1:
                 end = end_time
+
             midi_segment = create_midi_note_from_pitched_data(
                 start, end, pitched_data, word
             )