+++ ReleaseNotes.md
@@ -1,10 +1,12 @@
 # Version: 0.0.12
-Date: 2024.07.13
+Date: 2024.07.07
 - Changes:
+  - Reduce Memory usage by clearing cache in whisper
+  - Some bug fixes and improved error handling and logs
   - Fix hypen language download
 
 # Version: 0.0.11
-Date: 2024.06.02
+Date: 2024.07.06
 - Changes:
   - Better linebreak calculation
   - Remove cache folder when finished